use caller process's console now
